add: Add callbacks module to the vuln detector lib #2462
