+++ deps/wazuh_testing/wazuh_testing/vulnerability_detector.py
@@ -57,27 +57,6 @@ def mock_cve_db(func):
     return magic
 
 
-def callback_detect_vulnerability_scan_sleeping(line):
-    msg = rf"{VULNERABILITY_DETECTOR_PREFIX} Sleeping for (.*)..."
-    match = re.match(msg, line)
-
-    return match.group(1) if match is not None else ""
-
-
-def callback_detect_vulnerability_detector_disabled(line):
-    msg = rf"{VULNERABILITY_DETECTOR_PREFIX}DEBUG: Module disabled. Exiting..."
-    match = re.match(msg, line)
-
-    return match is not None
-
-
-def callback_detect_vulnerability_detector_enabled(line):
-    msg = r'(.*)wazuh-modulesd:vulnerability-detector(.*)'
-    match1 = re.match(msg, line)
-    msg = r'(.*)DEBUG: Module disabled. Exiting...(.*)'
-    match2 = re.match(msg, line)
-
-    return match1 is not None and match2 is None
 
 
 def make_vuln_callback(pattern, prefix=VULNERABILITY_DETECTOR_PREFIX):
